minor refractor
